requirements.txt
numpy
requests
tqdm
matplotlib
opencv-python
Pillow
torch
transformers
python-dotenv
# profanity_filter